requirements.txt
# Core dependencies
crawl4ai
python-dotenv
mcp
openai
supabase
langgraph
mem0ai
rouge-score
nltk
scikit-learn
numpy
psutil
toml
packaging
gitpython
pytest
pytest-asyncio
pytest-cov
requests
sentence-transformers
aiofiles

# hiRAG dependencies
networkx
umap-learn
numba
tiktoken
tenacity
tqdm
neo4j
community

# Optional dependencies
cohere
huggingface-hub
transformers
torch
accelerate
colpali-engine
pillow
bitsandbytes
aiofiles